WORKAROUND: Do not run validators for ZoneSerializer with `nested=True`
This is a workaround for NetBox issue #15351.

The new method of using standard serializers with the attribute `nested` set to `True`
instead of specific nested serializers results in the problem that the nested serialisers
created that way have validators attached to them, which causes problems at least in
combination with object custom fields.

This workaround restores the old nested serializer behaviour for `ZoneSerializer` until
the issue has been addressed in NetBox.